requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
certifi==2024.7.4
cffi==1.17.0
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.112.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.7
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0
